+++ plugin-api/src/main/java/org/sonar/api/measures/CoreMetrics.java
@@ -798,15 +798,7 @@ public final class CoreMetrics {
     .setOptimizedBestValue(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String BLOCKER_VIOLATIONS_KEY = "blocker_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> BLOCKER_VIOLATIONS = new Metric.Builder(BLOCKER_VIOLATIONS_KEY, "Blocker Issues", Metric.ValueType.INT)
     .setDescription("Blocker issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -816,15 +808,7 @@ public final class CoreMetrics {
     .setOptimizedBestValue(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String CRITICAL_VIOLATIONS_KEY = "critical_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> CRITICAL_VIOLATIONS = new Metric.Builder(CRITICAL_VIOLATIONS_KEY, "Critical Issues", Metric.ValueType.INT)
     .setDescription("Critical issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -834,15 +818,7 @@ public final class CoreMetrics {
     .setOptimizedBestValue(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String MAJOR_VIOLATIONS_KEY = "major_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> MAJOR_VIOLATIONS = new Metric.Builder(MAJOR_VIOLATIONS_KEY, "Major Issues", Metric.ValueType.INT)
     .setDescription("Major issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -852,15 +828,7 @@ public final class CoreMetrics {
     .setOptimizedBestValue(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String MINOR_VIOLATIONS_KEY = "minor_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> MINOR_VIOLATIONS = new Metric.Builder(MINOR_VIOLATIONS_KEY, "Minor Issues", Metric.ValueType.INT)
     .setDescription("Minor issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -870,15 +838,7 @@ public final class CoreMetrics {
     .setOptimizedBestValue(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String INFO_VIOLATIONS_KEY = "info_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> INFO_VIOLATIONS = new Metric.Builder(INFO_VIOLATIONS_KEY, "Info Issues", Metric.ValueType.INT)
     .setDescription("Info issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -899,15 +859,7 @@ public final class CoreMetrics {
     .setDeleteHistoricalData(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String NEW_BLOCKER_VIOLATIONS_KEY = "new_blocker_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_BLOCKER_VIOLATIONS = new Metric.Builder(NEW_BLOCKER_VIOLATIONS_KEY, "New Blocker Issues", Metric.ValueType.INT)
     .setDescription("New Blocker issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -918,15 +870,7 @@ public final class CoreMetrics {
     .setDeleteHistoricalData(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String NEW_CRITICAL_VIOLATIONS_KEY = "new_critical_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_CRITICAL_VIOLATIONS = new Metric.Builder(NEW_CRITICAL_VIOLATIONS_KEY, "New Critical Issues", Metric.ValueType.INT)
     .setDescription("New Critical issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -937,15 +881,7 @@ public final class CoreMetrics {
     .setDeleteHistoricalData(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String NEW_MAJOR_VIOLATIONS_KEY = "new_major_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_MAJOR_VIOLATIONS = new Metric.Builder(NEW_MAJOR_VIOLATIONS_KEY, "New Major Issues", Metric.ValueType.INT)
     .setDescription("New Major issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -956,15 +892,7 @@ public final class CoreMetrics {
     .setDeleteHistoricalData(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String NEW_MINOR_VIOLATIONS_KEY = "new_minor_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_MINOR_VIOLATIONS = new Metric.Builder(NEW_MINOR_VIOLATIONS_KEY, "New Minor Issues", Metric.ValueType.INT)
     .setDescription("New Minor issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -975,15 +903,7 @@ public final class CoreMetrics {
     .setDeleteHistoricalData(true)
     .create();
 
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final String NEW_INFO_VIOLATIONS_KEY = "new_info_violations";
-  /**
-   * @deprecated since 10.7 as {@link Severity} is deprecated. No replacement.
-   */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_INFO_VIOLATIONS = new Metric.Builder(NEW_INFO_VIOLATIONS_KEY, "New Info Issues", Metric.ValueType.INT)
     .setDescription("New Info issues")
     .setDirection(Metric.DIRECTION_WORST)
@@ -1329,18 +1249,14 @@ public final class CoreMetrics {
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #MAINTAINABILITY_ISSUES_KEY} instead.
    */
-  @Deprecated(since = "10.7")
   public static final String CODE_SMELLS_KEY = "code_smells";
 
   /**
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #MAINTAINABILITY_ISSUES} instead.
    */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> CODE_SMELLS = new Metric.Builder(CODE_SMELLS_KEY, "Code Smells", Metric.ValueType.INT)
     .setDescription("Code Smells")
     .setDirection(Metric.DIRECTION_WORST)
@@ -1354,18 +1270,14 @@ public final class CoreMetrics {
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #NEW_MAINTAINABILITY_ISSUES_KEY} instead.
    */
-  @Deprecated(since = "10.7")
   public static final String NEW_CODE_SMELLS_KEY = "new_code_smells";
 
   /**
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #NEW_MAINTAINABILITY_ISSUES} instead.
    */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_CODE_SMELLS = new Metric.Builder(NEW_CODE_SMELLS_KEY, "New Code Smells", Metric.ValueType.INT)
     .setDescription("New Code Smells")
     .setDirection(Metric.DIRECTION_WORST)
@@ -1380,18 +1292,14 @@ public final class CoreMetrics {
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #RELIABILITY_ISSUES_KEY} instead.
    */
-  @Deprecated(since = "10.7")
   public static final String BUGS_KEY = "bugs";
 
   /**
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #RELIABILITY_ISSUES} instead.
    */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> BUGS = new Metric.Builder(BUGS_KEY, "Bugs", Metric.ValueType.INT)
     .setDescription("Bugs")
     .setDirection(Metric.DIRECTION_WORST)
@@ -1405,18 +1313,14 @@ public final class CoreMetrics {
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #NEW_RELIABILITY_ISSUES_KEY} instead.
    */
-  @Deprecated(since = "10.7")
   public static final String NEW_BUGS_KEY = "new_bugs";
 
   /**
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #NEW_RELIABILITY_ISSUES} instead.
    */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_BUGS = new Metric.Builder(NEW_BUGS_KEY, "New Bugs", Metric.ValueType.INT)
     .setDescription("New Bugs")
     .setDirection(Metric.DIRECTION_WORST)
@@ -1431,18 +1335,14 @@ public final class CoreMetrics {
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #SECURITY_ISSUES_KEY} instead.
    */
-  @Deprecated(since = "10.7")
   public static final String VULNERABILITIES_KEY = "vulnerabilities";
 
   /**
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #SECURITY_ISSUES} instead.
    */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> VULNERABILITIES = new Metric.Builder(VULNERABILITIES_KEY, "Vulnerabilities", Metric.ValueType.INT)
     .setDescription("Vulnerabilities")
     .setDirection(Metric.DIRECTION_WORST)
@@ -1456,17 +1356,13 @@ public final class CoreMetrics {
    * SonarQube Quality Model
    *
    * @since 5.5
-   * @deprecated since 10.7. Use {@link #NEW_SECURITY_ISSUES_KEY} instead.
    */
-  @Deprecated(since = "10.7")
   public static final String NEW_VULNERABILITIES_KEY = "new_vulnerabilities";
 
   /**
    * SonarQube Quality Model
    *
-   * @deprecated since 10.7. Use {@link #NEW_SECURITY_ISSUES} instead.
    */
-  @Deprecated(since = "10.7")
   public static final Metric<Integer> NEW_VULNERABILITIES = new Metric.Builder(NEW_VULNERABILITIES_KEY, "New Vulnerabilities", Metric.ValueType.INT)
     .setDescription("New Vulnerabilities")
     .setDirection(Metric.DIRECTION_WORST)